all-train-data/schnet-delta/requirements.txt
pytorch==0.4.1
https://github.com/WardLT/schnetpack.git@add_inputs
numpy==1.15.4
